+++ apps/rehabstod/src/pages/CurrentSickLeaves/CurrentSickLeaves.tsx
@@ -8,8 +8,9 @@ import { TableHeader } from '../../components/Table/tableHeader/TableHeader'
 import { TableContentAlert } from '../../components/error/ErrorAlert/TableContentAlert'
 import { UnansweredCommunicationAlert } from '../../components/error/ErrorAlert/UnansweredCommunicationAlert'
 import { UserUrval } from '../../schemas'
-import { useGetSickLeavesFiltersQuery, useGetUserQuery, useLazyGetSickLeavesQuery } from '../../store/api'
+import { useGetUserQuery } from '../../store/api'
 import { useAppDispatch, useAppSelector } from '../../store/hooks'
+import { useGetSickLeavesFiltersQuery, useLazyGetSickLeavesQuery } from '../../store/sickLeaveApi'
 import { updateShowPersonalInformation } from '../../store/slices/settings.slice'
 import { resetSickLeaveFilters } from '../../store/slices/sickLeave.slice'
 import { SickLeaveColumn } from '../../store/slices/sickLeaveTableColumns.slice'
