+++ apps/rehabstod/src/pages/CurrentSickLeaves/components/ModifySicknessTableColumns.tsx
@@ -1,8 +1,9 @@
 import { useEffect } from 'react'
 import { ModifyTableColumns } from '../../../components/Table/ModifyTableColumns/ModifyTableColumns'
 import { filterTableColumns } from '../../../components/Table/utils/filterTableColumns'
-import { useGetSickLeavesFiltersQuery, useGetUserQuery } from '../../../store/api'
+import { useGetUserQuery } from '../../../store/api'
 import { useAppDispatch, useAppSelector, useUpdateUserPreferences } from '../../../store/hooks'
+import { useGetSickLeavesFiltersQuery } from '../../../store/sickLeaveApi'
 import { allSickLeaveColumns, sickLeaveColumnsString } from '../../../store/slices/sickLeaveTableColumns.selector'
 import { hideColumn, moveColumn, setColumnDefaults, showColumn } from '../../../store/slices/sickLeaveTableColumns.slice'
 import { isUserDoctor } from '../../../utils/isUserDoctor'
