+++ apps/rehabstod/src/pages/CurrentSickLeaves/components/TableBodyRows.tsx
@@ -13,8 +13,8 @@ import { TableRow } from '../../../components/Table/tableBody/TableRow'
 import { getEmptyFiltrationText, getEmptyTableText, getSearchText } from '../../../components/Table/utils/tableTextGeneratorUtils'
 import { User } from '../../../schemas'
 import { SickLeaveInfo } from '../../../schemas/sickLeaveSchema'
-import { useGetSickLeavesFiltersQuery } from '../../../store/api'
 import { useAppSelector } from '../../../store/hooks'
+import { useGetSickLeavesFiltersQuery } from '../../../store/sickLeaveApi'
 import { allSickLeaveColumns } from '../../../store/slices/sickLeaveTableColumns.selector'
 import { SickLeaveColumn } from '../../../store/slices/sickLeaveTableColumns.slice'
 import { getUnansweredCommunicationFormat } from '../../../utils/getUnansweredCommunicationFormat'
