+++ apps/rehabstod/src/pages/Home/components/OverviewStatistics.tsx
@@ -1,7 +1,9 @@
 import { IDSCard, IDSSpinner } from '@frontend/ids-react-ts'
+import { EmptyTableAlert } from '../../../components/Table/EmptyTableAlert'
 import { ErrorAlert } from '../../../components/error/ErrorAlert/ErrorAlert'
 import { UserUrval } from '../../../schemas'
-import { useGetSickLeavesSummaryQuery, useGetUserQuery } from '../../../store/api'
+import { useGetUserQuery } from '../../../store/api'
+import { useGetSickLeavesSummaryQuery } from '../../../store/sickLeaveApi'
 import { CountSickLeaveDegreesCard } from './card/CountSickLeaveDegreesCard'
 import { DiagnosisGroupsCard } from './card/DiagnosisGroupsCard'
 import { SickLeaveDegreesCard } from './card/SickLeaveDegreesCard'
@@ -9,7 +11,6 @@ import { SickLeaveLengthsCard } from './card/SickLeaveLengthsCard'
 import { StatisticsInformationCard } from './card/StatisticsInformationCard'
 import { GenderDivisionGraph } from './graph/GenderDivisionGraph'
 import { TotalSickLeavesGraph } from './graph/TotalSickLeavesGraph'
-import { EmptyTableAlert } from '../../../components/Table/EmptyTableAlert'
 
 export function OverviewStatistics() {
   const { data: user } = useGetUserQuery()
