+++ apps/rehabstod/src/pages/LUCertificates/LUCertificates.tsx
@@ -9,8 +9,9 @@ import { TableHeadingForUnit } from '../../components/Table/heading/TableHeading
 import { TableHeader } from '../../components/Table/tableHeader/TableHeader'
 import { filterHiddenColumns, filterTableColumns } from '../../components/Table/utils/filterTableColumns'
 import { useNavigateToStartPage } from '../../hooks/useNavigateToStartPage'
-import { useGetLUFiltersQuery, useGetUserQuery, useLazyGetLUCertificatesQuery } from '../../store/api'
+import { useGetUserQuery } from '../../store/api'
 import { useAppSelector } from '../../store/hooks'
+import { useGetLUFiltersQuery, useLazyGetLUCertificatesQuery } from '../../store/luApi'
 import { allLUUnitTableColumns } from '../../store/slices/luUnitTableColumns.selector'
 import { LUCertificatesColumn } from '../../store/slices/luUnitTableColumns.slice'
 import { updateShowPersonalInformation } from '../../store/slices/settings.slice'
