+++ apps/rehabstod/src/pages/LUCertificates/LUCertificatesFilters.tsx
@@ -9,8 +9,9 @@ import { SelectFilter } from '../../components/Table/filter/SelectFilter'
 import { TextSearchFilter } from '../../components/Table/filter/TextSearchFilter'
 import { getMultipleSelectPlaceholder } from '../../components/Table/filter/utils/getMultipleSelectPlaceholder'
 import { LUCertificatesFilter } from '../../schemas/luCertificatesSchema'
-import { useGetLUFiltersQuery, useGetUserQuery } from '../../store/api'
+import { useGetUserQuery } from '../../store/api'
 import { useAppSelector } from '../../store/hooks'
+import { useGetLUFiltersQuery } from '../../store/luApi'
 import { resetLUFilters, updateFilter } from '../../store/slices/luCertificates.slice'
 import { isUserDoctor } from '../../utils/isUserDoctor'
