+++ apps/rehabstod/src/pages/Patient/Patient.tsx
@@ -2,7 +2,7 @@ import { skipToken } from '@reduxjs/toolkit/query'
 import { useParams } from 'react-router-dom'
 import { PageContainer } from '../../components/PageContainer/PageContainer'
 import { StickyPortal } from '../../components/StickyContainer/StickyPortal'
-import { useGetPatientSickLeavesQuery } from '../../store/api'
+import { useGetPatientSickLeavesQuery } from '../../store/sickLeaveApi'
 import { OpenTabsDialog } from './components/OpenTabsDialog/OpenTabsDialog'
 import { PatientHeader } from './components/PatientHeader'
 import { PatientTabs } from './components/PatientTabs'
