+++ apps/rehabstod/src/pages/Patient/components/PatientRekoStatus.tsx
@@ -1,6 +1,6 @@
 import { useLocation } from 'react-router-dom'
 import { PatientSjukfall } from '../../../schemas/patientSchema'
-import { useGetSickLeavesFiltersQuery } from '../../../store/api'
+import { useGetSickLeavesFiltersQuery } from '../../../store/sickLeaveApi'
 import { SelectRekoStatus } from './SelectRekoStatus/SelectRekoStatus'
 
 export function PatientRekoStatus({
