+++ apps/rehabstod/src/pages/Patient/components/patientAG/PatientAGCertificatesTable.tsx
@@ -4,8 +4,9 @@ import { Table } from '../../../../components/Table/Table'
 import { TableHeader } from '../../../../components/Table/tableHeader/TableHeader'
 import { filterHiddenColumns, filterTableColumns } from '../../../../components/Table/utils/filterTableColumns'
 import { PatientTableError } from '../../../../components/error/ErrorAlert/PatientTableError'
-import { useGetAGCertificatesForPatientQuery, useGetUserQuery } from '../../../../store/api'
+import { useGetUserQuery } from '../../../../store/api'
 import { useAppSelector } from '../../../../store/hooks'
+import { useGetAGCertificatesForPatientQuery } from '../../../../store/sickLeaveApi'
 import { allPatientColumns } from '../../../../store/slices/patientTableColumns.selector'
 import { PatientColumn } from '../../../../store/slices/patientTableColumns.slice'
 import { isUserDoctor } from '../../../../utils/isUserDoctor'
