+++ apps/rehabstod/src/pages/Patient/components/patientSickLeaves/ModifyPatientTableColumns.tsx
@@ -1,8 +1,9 @@
 import { useEffect } from 'react'
 import { ModifyTableColumns } from '../../../../components/Table/ModifyTableColumns/ModifyTableColumns'
 import { filterTableColumns } from '../../../../components/Table/utils/filterTableColumns'
-import { useGetSickLeavesFiltersQuery, useGetUserQuery } from '../../../../store/api'
+import { useGetUserQuery } from '../../../../store/api'
 import { useAppDispatch, useAppSelector, useUpdateUserPreferences } from '../../../../store/hooks'
+import { useGetSickLeavesFiltersQuery } from '../../../../store/sickLeaveApi'
 import { allPatientColumns, patientColumnsString } from '../../../../store/slices/patientTableColumns.selector'
 import { PatientColumn, hideColumn, moveColumn, setColumnDefaults, showColumn } from '../../../../store/slices/patientTableColumns.slice'
 import { isUserDoctor } from '../../../../utils/isUserDoctor'
