+++ apps/rehabstod/src/pages/Patient/components/patientSickLeaves/PatientSickLeaves.tsx
@@ -2,16 +2,17 @@ import { skipToken } from '@reduxjs/toolkit/query'
 import { useParams } from 'react-router-dom'
 import { PageContainer } from '../../../../components/PageContainer/PageContainer'
 import { TableHeadingForUnit } from '../../../../components/Table/heading/TableHeadingForUnit'
-import { ModifyPatientTableColumns } from './ModifyPatientTableColumns'
-import { PatientSickLeavesTable } from './PatientSickLeavesTable'
-import { PatientOverview } from '../patientOverview/PatientOverview'
 import { PatientTableError } from '../../../../components/error/ErrorAlert/PatientTableError'
 import { PuResponse } from '../../../../schemas/patientSchema'
-import { useGetPatientSickLeavesQuery, useGetUserQuery } from '../../../../store/api'
+import { useGetUserQuery } from '../../../../store/api'
+import { useGetPatientSickLeavesQuery } from '../../../../store/sickLeaveApi'
 import { isDateBeforeToday } from '../../../../utils/isDateBeforeToday'
+import { isUserDoctor } from '../../../../utils/isUserDoctor'
 import { PatientRekoStatus } from '../PatientRekoStatus'
 import { PatientAGCertificatesTable } from '../patientAG/PatientAGCertificatesTable'
-import { isUserDoctor } from '../../../../utils/isUserDoctor'
+import { PatientOverview } from '../patientOverview/PatientOverview'
+import { ModifyPatientTableColumns } from './ModifyPatientTableColumns'
+import { PatientSickLeavesTable } from './PatientSickLeavesTable'
 
 export function PatientSickLeaves() {
   const { encryptedPatientId } = useParams()
