+++ apps/rehabstod/src/pages/Patient/components/patientSickLeaves/PatientTableBody.tsx
@@ -9,8 +9,8 @@ import { Tooltip } from '../../../../components/Tooltip/Tooltip'
 import { TooltipContent } from '../../../../components/Tooltip/TooltipContent'
 import { TooltipTrigger } from '../../../../components/Tooltip/TooltipTrigger'
 import { PatientSjukfallIntyg } from '../../../../schemas/patientSchema'
-import { useGetSickLeavesFiltersQuery } from '../../../../store/api'
 import { useAppSelector } from '../../../../store/hooks'
+import { useGetSickLeavesFiltersQuery } from '../../../../store/sickLeaveApi'
 import { allPatientColumns } from '../../../../store/slices/patientTableColumns.selector'
 import { PatientColumn } from '../../../../store/slices/patientTableColumns.slice'
 import { getCertificateColumnData } from '../../utils/getCertificateColumnData'
