+++ apps/rehabstod/src/store/luApi.ts
@@ -1,0 +1,46 @@
+import { DiagnosKapitel } from '../schemas/diagnosisSchema'
+import { Lakare } from '../schemas/lakareSchema'
+import { LUCertificatesFilter, LUCertificatesInfo } from '../schemas/luCertificatesSchema'
+import { api } from './api'
+
+const luApi = api.injectEndpoints({
+  endpoints: (builder) => ({
+    getLUCertificates: builder.query<LUCertificatesInfo, LUCertificatesFilter>({
+      query: (request) => ({
+        url: 'certificate/lu/unit',
+        method: 'POST',
+        body: request,
+      }),
+      async onQueryStarted(_, { dispatch }) {
+        dispatch(luApi.endpoints.getLUFilters.initiate(undefined, { forceRefetch: true }))
+      },
+    }),
+    getLUFilters: builder.query<
+      {
+        doctors: Lakare[]
+        allDiagnosisChapters: DiagnosKapitel[]
+      },
+      void
+    >({
+      keepUnusedDataFor: 0,
+      query: () => 'lu/filters',
+    }),
+    getPatientLUCertificates: builder.query<LUCertificatesInfo, { encryptedPatientId: string }>({
+      keepUnusedDataFor: 0,
+      query: (request) => ({
+        url: 'certificate/lu/person',
+        method: 'POST',
+        body: request,
+      }),
+    }),
+    getLUCertificatesDoctors: builder.query<{ doctors: Lakare[] }, void>({
+      query: () => ({
+        url: 'certificate/lu/doctors',
+        method: 'GET',
+      }),
+    }),
+  }),
+  overrideExisting: false,
+})
+
+export const { useGetLUFiltersQuery, useGetPatientLUCertificatesQuery, useLazyGetLUCertificatesQuery } = luApi
