+++ apps/rehabstod/src/store/sickLeaveApi.ts
@@ -1,0 +1,155 @@
+import { AGCertificatesInfo } from '../schemas/agCertificatesSchema'
+import { DiagnosKapitel } from '../schemas/diagnosisSchema'
+import { Lakare } from '../schemas/lakareSchema'
+import { Patient } from '../schemas/patientSchema'
+import {
+  OccupationType,
+  RekoStatusType,
+  SickLeaveFilter,
+  SickLeaveInfo,
+  SickLeaveSummary,
+  UnansweredCommunicationFilterType,
+} from '../schemas/sickLeaveSchema'
+import { api } from './api'
+
+const sickLeaveApi = api.injectEndpoints({
+  endpoints: (builder) => ({
+    getSickLeaves: builder.query<{ content: SickLeaveInfo[]; unansweredCommunicationError: boolean; srsError: boolean }, SickLeaveFilter>({
+      query: (request) => ({
+        url: 'sickleaves/active',
+        method: 'POST',
+        body: request,
+      }),
+      providesTags: ['User'],
+      async onQueryStarted(_, { dispatch }) {
+        dispatch(sickLeaveApi.endpoints.getSickLeavesFilters.initiate(undefined, { forceRefetch: true }))
+      },
+    }),
+    getSickLeavesFilters: builder.query<
+      {
+        activeDoctors: Lakare[]
+        allDiagnosisChapters: DiagnosKapitel[]
+        enabledDiagnosisChapters: DiagnosKapitel[]
+        nbrOfSickLeaves: number
+        hasOngoingSickLeaves: boolean
+        rekoStatusTypes: RekoStatusType[]
+        occupationTypes: OccupationType[]
+        unansweredCommunicationFilterTypes: UnansweredCommunicationFilterType[]
+        srsActivated: boolean
+      },
+      void
+    >({
+      query: () => ({
+        url: 'sickleaves/filters',
+      }),
+      providesTags: ['User', 'SickLeaves'],
+    }),
+    getSickLeavesSummary: builder.query<SickLeaveSummary, void>({
+      query: () => ({
+        url: 'sickleaves/summary',
+      }),
+      providesTags: ['User', 'SickLeaves'],
+    }),
+    getPatientSickLeaves: builder.query<Patient, { encryptedPatientId: string }>({
+      keepUnusedDataFor: 0,
+      query: (encryptedPatientId) => ({
+        url: 'sjukfall/patient',
+        method: 'POST',
+        body: encryptedPatientId,
+      }),
+      providesTags: ['Patient', 'SickLeaves'],
+    }),
+    getAGCertificatesForPatient: builder.query<AGCertificatesInfo, { encryptedPatientId: string }>({
+      keepUnusedDataFor: 0,
+      query: (request) => ({
+        url: 'certificate/ag/person',
+        method: 'POST',
+        body: request,
+      }),
+    }),
+    addVardenhet: builder.mutation<string[], { patientId: string; vardenhetId: string }>({
+      query: ({ patientId, vardenhetId }) => ({
+        url: 'sjukfall/patient/addVardenhet',
+        method: 'POST',
+        body: { patientId, vardenhetId },
+      }),
+      invalidatesTags: ['Patient'],
+    }),
+    addVardgivare: builder.mutation<string[], { patientId: string; vardgivareId: string }>({
+      query: ({ patientId, vardgivareId }) => ({
+        url: 'sjukfall/patient/addVardgivare',
+        method: 'POST',
+        body: { patientId, vardgivareId },
+      }),
+      invalidatesTags: ['Patient'],
+    }),
+    setRekoStatus: builder.mutation<
+      void,
+      { patientId: string; status: RekoStatusType; sickLeaveTimestamp: string; filter: SickLeaveFilter }
+    >({
+      query: ({ patientId, status, sickLeaveTimestamp }) => ({
+        url: 'reko',
+        method: 'POST',
+        body: { patientId, statusId: status.id, sickLeaveTimestamp },
+      }),
+      async onQueryStarted({ patientId, status, filter }, { dispatch, queryFulfilled }) {
+        dispatch(
+          sickLeaveApi.util.updateQueryData('getSickLeaves', filter, (draft) => {
+            const index = draft.content.findIndex((sickLeave) => sickLeave.patient.id === patientId)
+            if (index !== -1) {
+              // eslint-disable-next-line no-param-reassign
+              draft.content[index].rekoStatus = { status }
+            }
+          })
+        )
+        try {
+          await queryFulfilled
+        } catch {
+          dispatch(api.util.invalidateTags(['User']))
+        }
+      },
+    }),
+    giveSjfConsent: builder.mutation<
+      { registeredBy: string; responseCode: string; responseMessage: string },
+      { days: number; onlyCurrentUser: boolean; patientId: string; encryptedPatientId: string }
+    >({
+      query: ({ days, onlyCurrentUser, patientId }) => ({
+        url: 'consent',
+        method: 'POST',
+        body: { days, onlyCurrentUser, patientId },
+      }),
+      async onQueryStarted({ encryptedPatientId }, { dispatch, queryFulfilled }) {
+        try {
+          const {
+            data: { responseCode },
+          } = await queryFulfilled
+          dispatch(
+            sickLeaveApi.util.updateQueryData('getPatientSickLeaves', { encryptedPatientId }, (draft) =>
+              Object.assign(draft, {
+                sjfMetaData: {
+                  ...(draft.sjfMetaData ?? {}),
+                  samtyckeFinns: responseCode === 'OK',
+                },
+              })
+            )
+          )
+        } catch {
+          dispatch(api.util.invalidateTags(['Patient']))
+        }
+      },
+    }),
+  }),
+  overrideExisting: false,
+})
+
+export const {
+  useAddVardenhetMutation,
+  useAddVardgivareMutation,
+  useGetAGCertificatesForPatientQuery,
+  useGetPatientSickLeavesQuery,
+  useGetSickLeavesFiltersQuery,
+  useGetSickLeavesSummaryQuery,
+  useGiveSjfConsentMutation,
+  useLazyGetSickLeavesQuery,
+  useSetRekoStatusMutation,
+} = sickLeaveApi
